Handle EINTR in Oz instead of C++.
This is required mainly because of timeslicing within a system thread.
The C++ approach is loop until an EINTR is not seen, but this will block
the system thread and Mozart has no way to interrupt because C++ code is
outside of the VM's reach.

In this commit, the EINTR error code will be considered "succeed", with
a return value indicating whether EINTR happened, and the functions will
be repeated *in Oz* until the EINTR no longer exists. This allows the VM
steps in and timeslice to other threads.

(Perhaps we could directly use exception handling?)