Assign a huge HWM to syncpub to significantly reduce drop rate.
Without the HWM, the drop rate was like 20k out of 500k messages.
Imposing the HWM in the same order of the 500k messages reduces the
dropped message to 3 or less.

(We could do better. erlzmq2 doesn't drop any messages as far as I can
see, but I can't make erlzmq2 to work with 3.1.1...)